unified all setUps methods to ApplicationTests class
