+++ src/test/java/br/ufpb/dcx/apps4society/educapi/EducApiApplicationTests.java
@@ -1,15 +1,19 @@
 package br.ufpb.dcx.apps4society.educapi;
 
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.ComponentScan;
+import org.junit.jupiter.api.BeforeAll;
 
-@SpringBootTest
-@ComponentScan(basePackages = "repositories")
+import static io.restassured.RestAssured.*;
+
 public class EducApiApplicationTests {
 
-	@Test
-	void contextLoads() {
-	}
+	public static String invalidToken = "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJqb3NlMTdAZWR1Y2FwaS5jb20iLCJleHAiOjE2ODA2OTc2MjN9." +
+			"qfwlZuirBvosD82v-7lHxb8qhH54_KXR20_0z3guG9rZOW68l5y3gZtvugBtpevmlgK76dsa4hOUPOooRiJ3ng";
 
+	@BeforeAll
+	public static void setUp(){
+
+		baseURI = "http://localhost";
+		port = 8080;
+		basePath = "/v1/api/";
+	}
 }
