+++ src/test/java/br/ufpb/dcx/apps4society/educapi/resources/UserResourceIntegrationTest.java
@@ -1,28 +1,17 @@
 package br.ufpb.dcx.apps4society.educapi.resources;
 
+import br.ufpb.dcx.apps4society.educapi.EducApiApplicationTests;
 import br.ufpb.dcx.apps4society.educapi.utils.FileUtils;
 import br.ufpb.dcx.apps4society.educapi.utils.USER_RequestsUtil;
 import io.restassured.http.ContentType;
 import io.restassured.response.*;
 import org.json.JSONObject;
 import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import static io.restassured.RestAssured.*;
 
-public class UserResourceIntegrationTest {
-
-    private static String invalidToken = "eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJqb3NlMTdAZWR1Y2FwaS5jb20iLCJleHAiOjE2ODA2OTc2MjN9." +
-            "qfwlZuirBvosD82v-7lHxb8qhH54_KXR20_0z3guG9rZOW68l5y3gZtvugBtpevmlgK76dsa4hOUPOooRiJ3ng";
-
-    @BeforeAll
-    public static void setUp(){
-
-        baseURI = "http://localhost";
-        port = 8080;
-        basePath = "/v1/api/";
-    }
+public class UserResourceIntegrationTest extends EducApiApplicationTests {
 
     @Test
     public void insertUserByNameEmailPassword_shouldReturn201Test() throws Exception {
